+++ src/cybergarage/upnp/ActionList.cpp
@@ -28,9 +28,9 @@ const char *CyberLink::ActionList::ELEM_NAME = "actionList";
 // Methods
 ////////////////////////////////////////////////
 
- void ActionList::clear()  {
-  int nAct = size();
-  for (int n = 0; n < nAct; n++) {
+void ActionList::clear()  {
+  size_t nAct = size();
+  for (size_t n = 0; n < nAct; n++) {
     Action *act = getAction(n);
     delete act;
   }
